rename and add license header
