--- test/plugin/scenarios/jetty-scenario/jettyclient-scenario/src/main/java/test/apache/skywalking/testcase/jettyclient/Application.java
@@ -1,21 +1,0 @@
-package test.apache.skywalking.testcase.jettyclient;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.boot.web.support.SpringBootServletInitializer;
-import org.springframework.context.annotation.ComponentScan;
-
-@EnableAutoConfiguration
-@SpringBootApplication
-@ComponentScan("test.apache.skywalking.testcase.jettyclient")
-public class Application extends SpringBootServletInitializer {
-
-    public static void main(String[] args) {
-        try {
-            SpringApplication.run(Application.class, args);
-        } catch (Exception e) {
-            // Never do this
-        }
-    }
-}
