+++ test/plugin/scenarios/jetty-scenario/jettyclient-scenario/src/main/resources/log4j2.xml
@@ -1,4 +1,21 @@
 <?xml version="1.0" encoding="UTF-8"?>
+<!--
+  ~ Licensed to the Apache Software Foundation (ASF) under one or more
+  ~ contributor license agreements.  See the NOTICE file distributed with
+  ~ this work for additional information regarding copyright ownership.
+  ~ The ASF licenses this file to You under the Apache License, Version 2.0
+  ~ (the "License"); you may not use this file except in compliance with
+  ~ the License.  You may obtain a copy of the License at
+  ~
+  ~     http://www.apache.org/licenses/LICENSE-2.0
+  ~
+  ~ Unless required by applicable law or agreed to in writing, software
+  ~ distributed under the License is distributed on an "AS IS" BASIS,
+  ~ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+  ~ See the License for the specific language governing permissions and
+  ~ limitations under the License.
+  ~
+  -->
 <Configuration status="debug">
     <Appenders>
         <Console name="Console" target="SYSTEM_OUT">
@@ -6,11 +23,6 @@
         </Console>
     </Appenders>
     <Loggers>
-        <logger name="com.a.eye.skywalking.ui" level="debug" additivity="false">
-            <AppenderRef ref="Console"/>
-        </logger>
-        <!--<logger name="org.skywalking.apm.dependencies" level="off">-->
-        <!--</logger>-->
         <Root level="debug">
             <AppenderRef ref="Console"/>
         </Root>
