+++ test/plugin/scenarios/jetty-scenario/jettyserver-scenario/src/main/java/org/apache/skywalking/apm/testcase/jettyserver/Application.java
@@ -1,0 +1,37 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ */
+package org.apache.skywalking.apm.testcase.jettyserver;
+
+import java.net.InetSocketAddress;
+import org.apache.skywalking.apm.testcase.jettyserver.servlet.CaseServlet;
+import org.eclipse.jetty.server.Server;
+import org.eclipse.jetty.servlet.ServletContextHandler;
+
+public class Application {
+
+    public static void main(String[] args) throws Exception {
+        Server jettyServer = new Server(new InetSocketAddress("0.0.0.0",
+            Integer.valueOf(18080)));
+        String contextPath = "/jettyserver-case";
+        ServletContextHandler servletContextHandler = new ServletContextHandler(ServletContextHandler.NO_SESSIONS);
+        servletContextHandler.setContextPath(contextPath);
+        servletContextHandler.addServlet(CaseServlet.class, CaseServlet.SERVLET_PATH);
+        jettyServer.setHandler(servletContextHandler);
+        jettyServer.start();
+    }
+}
